filter data perbaikan

+++ app/Http/Controllers/UserController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use App\Role;
 use App\User;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Validator;
 
 
@@ -19,7 +20,7 @@ class UserController extends Controller
     {
         $users = User::all();
 
-        return view('user-all',[
+        return view('user-all', [
             'users' => $users
         ]);
     }
@@ -41,21 +42,25 @@ class UserController extends Controller
     /**
      * Store a newly created resource in storage.
      *
-     * @param  \Illuminate\Http\Request  $request
+     * @param  \Illuminate\Http\Request $request
      * @return \Illuminate\Http\Response
      */
     public function store(Request $request)
     {
-        $this->validate($request, [
+        $validator = Validator::make($request->all(), [
             'nama' => 'required',
             'username' => 'required | unique:users',
-            'password' => 'required',
+            'password' => 'required | min:6',
         ], [
             'required' => 'kolom di tidak boleh kosong',
-            'unique' => 'username tersebut telah digunakan'
+            'unique' => 'username tersebut telah digunakan',
+            'min' => 'password minimal berjumlah 6 karakter'
         ]);
 
-        $validator = Validator::make();
+        if ($validator->fails()) {
+            return redirect()->route('user.create')->withInput()
+                ->withErrors($validator);
+        }
 
         User::create([
             'nama' => $request->nama,
@@ -70,8 +75,8 @@ class UserController extends Controller
     /**
      * Display the specified resource.
      *
-     * @param  int  $id
-x     * @return \Illuminate\Http\Response
+     * @param  int $id
+     * x     * @return \Illuminate\Http\Response
      */
     public function show($id)
     {
@@ -81,7 +86,7 @@ x     * @return \Illuminate\Http\Response
     /**
      * Show the form for editing the specified resource.
      *
-     * @param  int  $id
+     * @param  int $id
      * @return \Illuminate\Http\Response
      */
     public function edit($id)
@@ -97,8 +102,8 @@ x     * @return \Illuminate\Http\Response
     /**
      * Update the specified resource in storage.
      *
-     * @param  \Illuminate\Http\Request  $request
-     * @param  int  $id
+     * @param  \Illuminate\Http\Request $request
+     * @param  int $id
      * @return \Illuminate\Http\Response
      */
     public function update(Request $request, $id)
@@ -106,35 +111,63 @@ x     * @return \Illuminate\Http\Response
         $this->validate($request, [
             'nama' => 'required',
             'username' => 'required|unique:users',
-            'password' => 'required|string|min:6|confirmed',
         ], [
             'required' => 'kolom di tidak boleh kosong',
             'unique' => 'username tersebut telah digunakan',
-            'min' => 'Kolom di atas setidaknya mengandung :digits karakter',
-            'confirmed' => 'Password yang anda masukkan tidak sama'
         ]);
 
         User::find($id)->update([
             'nama' => $request->nama,
             'username' => $request->username,
-            'password' => bcrypt($request->password),
             'role_id' => $request->role_id
         ]);
 
-        return redirect()->route('user.index')->with('edited', 'Data user "'.$request->nama.'" berhasil diperbarui');
+        return redirect()->route('user.index')->with('edited', 'Data user "' . $request->nama . '" berhasil diperbarui');
     }
 
     /**
      * Remove the specified resource from storage.
      *
-     * @param  int  $id
+     * @param  int $id
      * @return \Illuminate\Http\Response
      */
     public function destroy($id)
     {
-        $user = User::find($id);
-        User::find($id)->delete();
+        if (Auth::user()->id != $id) {
+            $user = User::find($id);
+            User::find($id)->delete();
 
-        return redirect()->route('user.index')->with('deleted', 'User "'.$user->nama.'" berhasil dihapus.');
+            return redirect()->route('user.index')->with('deleted', 'User "' . $user->nama . '" berhasil dihapus.');
+        } else {
+            return redirect()->route('user.index')->with('deleted', 'User tersebut sedang digunakan!');
+        }
+
+    }
+
+
+    public function gantiPassword(Request $request, $id)
+    {
+        $this->validate($request, [
+            'password' => 'required|string|min:6|confirmed',
+        ], [
+            'required' => 'kolom di tidak boleh kosong',
+            'min' => 'Password setidaknya mengandung 6 karakter',
+            'confirmed' => 'Password yang anda masukkan tidak sama'
+        ]);
+
+        User::find($id)->update([
+            'password' => bcrypt($request->password),
+        ]);
+
+        return redirect()->route('user.index')->with('edited', 'Password berhasil diperbarui');
+    }
+
+    public function gantiRole(Request $request, $id)
+    {
+        User::find($id)->update([
+            'role_id' => $request->role_id
+        ]);
+
+        return redirect()->route('user.index')->with('edited', 'Role berhasil diperbarui');
     }
 }
